requirements.txt
asgiref==3.6.0
asttokens==2.2.1
backcall==0.2.0
black==23.1.0
click==8.1.3
crispy-bootstrap4==2022.1
decorator==5.1.1
Django==4.1.7
django-crispy-forms==2.0
djangorestframework==3.14.0
et-xmlfile==1.1.0
executing==1.2.0
ipdb==0.13.13
ipython==8.11.0
jedi==0.18.2
matplotlib-inline==0.1.6
mypy-extensions==1.0.0
numpy==1.24.2
openpyxl==3.1.2
packaging==23.0
pandas==1.5.3
parso==0.8.3
pathspec==0.11.1
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.1.1
prompt-toolkit==3.0.38
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.14.0
python-dateutil==2.8.2
pytz==2022.7.1
six==1.16.0
sqlparse==0.4.3
stack-data==0.6.2
tomli==2.0.1
traitlets==5.9.0
wcwidth==0.2.6
